Refactor into separate directories
