+++ src/SimulationService.jl
@@ -14,7 +14,7 @@ import DataFrames: DataFrame
 import HTTP: Request, Response
 import JobSchedulers: scheduler_start, set_scheduler, scheduler_stop, submit!, job_query, result, generate_id, update_queue!, Job, JobSchedulers
 
-include("./SciMLInterface.jl"); import .SciMLInterface: sciml_operations, use_operation, conversions_for_valid_inputs
+include("./contracts/Interface.jl"); import .Interface: sciml_operations, use_operation, conversions_for_valid_inputs
 include("./service/Service.jl"); import .Service.ArgIO: prepare_output, prepare_input
 include("./Settings.jl"); import .Settings: settings
