+++ src/service/Service.jl
@@ -5,6 +5,7 @@ include("../Settings.jl")
 
 include("./AssetManager.jl")
 include("./ArgIO.jl")
-@reexport import .ArgIO, .AssetManager
+include("./Queuing.jl")
+@reexport import .ArgIO, .AssetManager, .Queuing
 
 end # module Service
