+++ test/sciml.jl
@@ -1,4 +1,4 @@
-using SimulationService, AlgebraicPetri, DataFrames, DifferentialEquations, ModelingToolkit, Symbolics, EasyModelAnalysis, Catlab, Catlab.CategoricalAlgebra, JSON3, UnPack, SimulationService.SciMLInterface.SciMLOperations
+using SimulationService, AlgebraicPetri, DataFrames, DifferentialEquations, ModelingToolkit, Symbolics, EasyModelAnalysis, Catlab, Catlab.CategoricalAlgebra, JSON3, UnPack, SimulationService.Interface.Operations
 using CSV, DataFrames, JSONTables
 using ForwardDiff
 
@@ -29,12 +29,12 @@ j = JSON3.write(body)
 forecast_fn = _log("forecast.json")
 write(forecast_fn, j)
 
-df = SimulationService.SciMLInterface.simulate(; nt...)
+df = SimulationService.Interface.simulate(; nt...)
 @test df isa DataFrame
 
 params["t1"] = 0.1
 nt = (; model = petri, params, initials, tspan)
-df2 = SimulationService.SciMLInterface.simulate(; nt...)
+df2 = SimulationService.Interface.simulate(; nt...)
 
 timesteps = df.timestamp
 data = Dict(["Susceptible" => df[:, 2]])
